feat: statisticController 없애고 planController로 연결
