+++ src/routes/plans.js
@@ -8,6 +8,7 @@ const {
   editPlan,
   deletePlan,
   notifyTodayPlan,
+  getStatistics,
 } = require('../controllers/planController.js');
 
 router.get('/', ensureAuthorization, getPlans);
@@ -15,5 +16,6 @@ router.post('/', ensureAuthorization, addPlan);
 router.put('/:id', ensureAuthorization, editPlan);
 router.delete('/:id', ensureAuthorization, deletePlan);
 router.get('/notifications/today', ensureAuthorization, notifyTodayPlan);
+router.get('/statistics/week', ensureAuthorization, getStatistics);
 
 module.exports = router;
